Refactor module visibility
