+++ rng_tools/src/generators/gen3/mod.rs
@@ -1,10 +1,11 @@
 mod egg;
 mod feebas_sid;
-pub mod seed3;
+mod seed;
 mod stationary;
 mod tidsid;
 
 pub use egg::*;
 pub use feebas_sid::*;
+pub use seed::*;
 pub use stationary::*;
 pub use tidsid::*;
