+++ rng_tools/src/generators/gen3/seed.rs
