+++ rng_tools/src/generators/gen4/mod.rs
@@ -1,6 +1,7 @@
 mod id;
-pub mod seed4;
+mod seed;
 mod seed_time;
 
 pub use id::*;
+pub use seed::*;
 pub use seed_time::*;
