+++ rng_tools/src/generators/gen6/seed.rs
