+++ rng_tools/src/generators/mod.rs
@@ -1,9 +1,4 @@
-mod gen2;
-mod gen3;
-mod gen4;
-mod gen6;
-
-pub use gen2::*;
-pub use gen3::*;
-pub use gen4::*;
-pub use gen6::*;
+pub mod gen2;
+pub mod gen3;
+pub mod gen4;
+pub mod gen6;
